libs/core/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cahoots-core"
version = "0.1.0"
description = "Core infrastructure for Cahoots"
requires-python = ">=3.11"
dependencies = [
    # Core dependencies
    "fastapi>=0.104.1",
    "fastapi-mail>=1.4.2",
    "uvicorn>=0.24.0",
    "pydantic>=2.5.2",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "email-validator>=2.1.0",
    "structlog>=23.2.0",
    
    # Database
    "sqlalchemy>=2.0.23",
    "alembic>=1.12.1",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",
    "redis>=5.0.1",
    
    # AI/ML
    "openai>=1.0.0",
    "anthropic>=0.5.0",
    "tiktoken>=0.5.0",
    
    # Monitoring and metrics
    "prometheus-client>=0.19.0",
    "opentelemetry-api>=1.21.0",
    "opentelemetry-sdk>=1.21.0",
    "opentelemetry-instrumentation-fastapi>=0.42b0",
    "sentry-sdk>=1.39.1",
    "psutil>=5.9.6",
    
    # Task management
    "trello>=0.9.7",
    "github3.py>=4.0.1",
    "PyGithub>=2.1.1",
    
    # Billing and subscriptions
    "stripe>=7.6.0",
    
    # Security
    "cryptography>=41.0.5",
    "authlib>=1.2.1",
    
    # Development tools
    "black>=23.11.0",
    "isort>=5.12.0",
    "mypy>=1.7.1",
    "pre-commit>=3.5.0",
    
    # Deployment
    "gunicorn>=21.2.0",
    "supervisor>=4.2.5",
    "docker>=6.1.3",
    "kubernetes>=28.1.0",
    
    # Additional dependencies
    "jsonschema>=4.20.0",
    "anyio>=3.7.1",
    "bcrypt>=4.0.1",
    "click>=8.1.7",
    "ddtrace>=2.18.1",
    "greenlet>=3.0.1",
    "h11>=0.14.0",
    "httptools>=0.6.1",
    "mypy-extensions>=1.0.0",
    "packaging>=23.2",
    "pathspec>=0.11.2",
    "platformdirs>=3.11.0",
    "PyYAML>=6.0.1",
    "watchfiles>=0.21.0",
    "websockets>=12.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "coverage>=7.3.2",
    "pytest-mock>=3.11.0",
    "httpx>=0.24.0",
    "pytest-randomly>=3.15.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/cahoots_core"]

[tool.hatch.build]
packages = ["src/cahoots_core"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=src --cov-report=term-missing"
asyncio_mode = "strict"
pythonpath = ["src"] 